[#66753] Rework ng GridAreaService using Signals
Refactoring with Signals, Angular's latest take on reactivity, is an
attempt to make `GridComponent` subclasses (and other `GridAreaService`
consumers) work properly with `ChangeDetectionStrategy.OnPush`.

Specifically, this patch should address an issue where the UI is not
updated in a timely fashion after users remove a Widget from the
Projects Overview page.