[#67244] Remove Dashboards module
Also removes ng `DashboardComponent` and related code.

This functionality has been disabled for a long period of time and it is
not envisaged that it will be re-enabled in the future. Instead,
Dashboards will be reimplemented in the Overviews module.

See opf/openproject#20045
See https://community.openproject.org/work_packages/66904