Fix flaky meeting spec partially
failing spec: `modules/meeting/spec/features/structured_meetings/structured_meeting_update_flash_spec.rb`

When run locally, it does not wait enough after saving an agenda item
and it sometimes creates synchronizations issues such as:

- click on the menu icon of an agenda item too early, then the items are
  updated by the turbo response and the menu is closed, so it fails to
  find the ".Overlay" element. It is retried so it's not that big of a
  problem, but it's not ideal as it waits for nothing.

- save an agenda item on a browser window, then switch to another tab
  and expect the update notification to be displayed. Well sometimes
  it's not saved yet, so the notification is not displayed, and the test
  fails.

- A variant of the previous one: save an agenda item on a browser
  window, then switch to another tab and reload the page. Sometimes the
  save is not finished yet, so the page reloads but the agenda item is
  not there yet.

The fix is to wait for the save to complete. Either by waiting for the
record to be created or updated, and then for the element to be
displayed. Or also by just sleeping 1 second because I could not find a
better synchronization mechanism.

Note that this still not fixes the flakiness where the `travel_to` seems
to not be effective anymore when the recurring meeting is created. See run
https://github.com/opf/openproject/actions/runs/17466279904/job/49602727897
for instance.