Implement MutationHelper, extracting boilerplate
Moves most of the logic out of the Stimulus Controller and into the new
helper. Adds a small debounce to coalesce consecutive muation updates.

`MutationHelper` is conceived as aa declarative wrapper around
`MutationObserver` - however it remains to be seen how useful it will be
in other DOM change use cases.