Sort hierarchy items by the name they are displayed with
The lists were a kind of make-shift solution to allow
using hierarchy custom fields in all of our UIs. However,
the sort order coming out of the service was rather useless
to a user. By sorting by the ancestry path, we sort the list by
the string that's later visible to the user, thus making
it easier to navigate the corresponding dropdown.