Confirm provider creation to user at end of OIDC provider wizard
So far we didn't redirect the user anywhere and they couldn't know
that they successfully created their provider. Now we redirect them
to the show view and display a success message.

As part of this change the "edit mode" was renamed to "new mode", because
it's exclusively set to true during initial provider creation, so it's rather
reflecting on being tied to the "new" user flow, than to any "edit" user flow.