Redirect user to show view after creating SAML provider
This allows them to see the sidebar that's regularly part of the
show view and only hidden during the initial creation of a provider.

For that we needed to change the show action to not render a turbostream
response anymore. This behaviour is now consistent with the OpenID Connect
provider controller again.