Fix flaky spec structured_meeting_crud_spec.rb
failing spec: modules/meeting/spec/features/structured_meetings/structured_meeting_crud_spec.rb:287
failing run: https://github.com/opf/openproject/actions/runs/18215876372/job/51864889174

There were multiple failures locally, and multiple causes:

First one, the dialog "Meeting participant" was doing
`MeetingParticipant.model_name.human`, but no translation existed for
:en, so it tried the fallback locales, which means all supported
locales. The i18n translations are lazily loaded in tests for
performance reasons, so initialization has to be done. It can take more
than 3 seconds on a loaded system, which means that the expectation
fails because the backend has not rendered the modal yet as it is still
busy loading the translations for _all_ languages.

Fix was to add a translation for
`activerecord.models.meeting_participant` key.

Second is that the interaction with some modals is done too early: the
modal has still not finished its appearing animation (it zooms in), so
some capybara actions do not quite work: for instance when capybara
tries to click in a field, it capture the position of the element,
locate the center, and clicks on that point. But as the animation is in
progress, the element being clicked has already moved and Capybara
clicks on the wrong place.

Fix was to modify `expect_modal` helper to always wait for this
animation to complete using `wait_for_size_animation_completion` helper.

Third one was a race condition when opening the meeting participant list
after copying and starting a meeting. When checking if all participants
are listed as expected, it also checks for the presence of the "Mark as
attended" button for each participant if the meeting is in progress.
Problem is: after clicking "Start meeting", the test opens the "manage
participants" modal immediately. If the meeting status has not finished
updating yet, the "manage participants" modal will be rendered without
the "Mark as attended" buttons, and when the participant presence is
checked, the meeting status has been updated in the meantime. So it
checks for the buttons presence, but they are not => failure.

Fix was to wait for the meeting status to be updated after starting the
meeting.