Add rake db:seed:performance
This is an extended seed task that seeds
many users, projects and work packages at the
end of the regular seeding.

Due to the extensive runtime of the seeder, it has
been designed to be interruptible. This means it will commit
changes more often to the database and every fully seeded project
will be immediately available.