Fix flaky spec
failing spec: ./spec/features/work_packages/details/date_editor_spec.rb:481
failing run: https://github.com/opf/openproject/actions/runs/18496811091/job/52702962599

When the working days only checkbox is toggled, the preview is triggered and updates the datepicker modal.

There was a race condition: if switching to automatic scheduling mode
before the preview finished, the preview would update the datepicker
modal and switching to automatic would be ignored (maybe it should not
be ignored). This was causing the spec to fail because the "working days
only" checkbox was not disabled.

Fix is to wait for the preview to finish before switching to automatic
scheduling mode.