Only count total when necessary
It's easily possible that the selects do not
require the calculation of the total number of elements.
However, since we resolve the number of total elements ahead
of constructing the complex SQL query, there's no way for the
database to optimize this out. Thus we have to manually skip
the total calculation if we do not need to perform it.