[68560] Fix change detection in update ancestors service
Fixes https://community.openproject.org/wp/68560

The problem was that the `WorkPacakges::UpdateService` updates the
related work packages in two phases:

- First it updates the ancestors, including itself. This updates work,
  remaining work, percent complete, switching to automatic mode, and
  ignore non-working days attributes.

- Second, if there are changes to dates, it updates the work packages to
  reschedule: the ancestors and successors. This updates the start date,
  finish date, and duration attributes.

This means that if something is changed in the first phase, the
initiating work package will be saved and there will be no changes left
to detect that a rescheduling of ancestors is needed too.

When having work packages with automatically generated subjects, that's
what happens: the subject is always updated to a placeholder text
("Automatically generated through type <type name>"), leading to the
`UpdateAncestorsService` always detecting a change and so always saving
it. This inhibits the rescheduling of ancestors.

The fix is to have a stricter change detection in the
`UpdateAncestorsService` to avoid saving if nothing was changed by the
service itself.

It feels brittle to me, and it's probable that modifying some attributes
along with dates could lead to the same problem. For instance modifying
work (estimated hours) and dates at the same time using the REST API.