Rework Repositories Settings form using Stimulus
Replaces ad-hoc (previously jQuery) code with `disable-when-checked` and
`show-when-checked` controllers.

Also adds specific handling for select fields to `disable-when-checked`
controller.

N.B. this introduces a behaviour change: hiding repository checkout
fields rather than toggling their `required `field. THis is to align
closer to our Primer-implemented forms, which do not rely on client-side
validations.