hack to resolve base tag conflicting with openapi explorer by overriding replaceState (+ using pushState instead)
openapi explorer uses replaceState (pushState in later versions) with
anchor only, but this doesn't work with base tag, as anchor is added to
the base url instead of current url. So override the replaceState
function to construct the full url based on current location. Also use
pushState instead to remember the history