Mark user's name property as read-only
This property was inherited from the PrincipalRepresenter where
it's writable. For users this would lead to a server error, since
there is no name attribute on the user model.

We could've just marked the property as `writable: false`, though
then changes to it would just be ignored without any API feedback,
but marking it as read-only through the contract is not possible, since
the contract can only validate attributes that really do exist.

To be able to provide API feedback, I overwrote the setter to raise a proper
validation error.