Try to fix a flaky spec
spec is spec/features/projects/project_custom_fields/overview_page/dialog/inputs_spec.rb:702
run is https://github.com/opf/openproject/actions/runs/19696080834/job/56421276049?pr=21163

Not entirely sure what's going on, but on the screenshot the field looks
active, so it's clicked, but for some reason the options are not
visible.

Try to work around the issue by waiting for the options to be visible,
and click the input again if not. Check every 0.25 seconds and try for 3
seconds before giving up.