Do not break out of Turbo Frames if nothing rendered
Turbo should do nothing if it encounters a completely empty response.
As such this commit completely removes the layout from Widget Controller
responses, explicitly setting HTTP response status to 204 No Content for
good measure. This should prevent Turbo from "breaking out" of the Turbo
Frame (redirecting to a blank page) when a widget is not rendered.